Create "Player" object
Creates an object that will represent the player.
The controllers will drive the player's position,
interaction will stem from the player, etc.

We are not using the HMD/Rig as a traditional VR set-up would,
since this is not a VR app per se; it's an XR setup for the Igloo
system. So we need bespoke controls, interaction, etc.

Eventually, interaction rays will extend from the player, etc.

Eventually, the Igloo display system will track around the player's
position.