Add Igloo cams to scene
Add Igloo cams to the scene. It should now render out to
the Igloo. Untested because we haven't set it up there yet.

We use our own Igloo prefab with the cursor disabled and the
height modified so that the camera centers around our Player object.

We choose not to use Igloo's Player capabilities because
(1) It appears broken on first glance; likely there's a fix,
but we have our own player system already. It was spinning around
uncontrolably.
(2) We have already started developing our own Player System.
I think we want to use our own to get more control for novel
interaction design; Igloo's approach seemed plug-and-play but
inflexible. We can always incorporate some of their systems
for e.g. UI later.